OPENEUROPA-3387: Port render test to browser test.
